refactor: ProjectProcess项目变更后置处理：新增删除项目一并删除项目配置

+++ arius-admin/arius-admin-core/src/main/java/com/didichuxing/datachannel/arius/admin/core/service/app/impl/ProjectConfigServiceImpl.java
@@ -112,11 +112,13 @@ public class ProjectConfigServiceImpl implements ProjectConfigService {
 	
 	/**
 	 * @param projectId
-	 * @return
 	 */
 	@Override
-	public Integer deleteByProjectId(int projectId) {
-		return projectConfigDAO.deleteByProjectId(projectId);
+	@Transactional(rollbackFor = Exception.class)
+	public void deleteByProjectId(int projectId) {
+		if (projectConfigDAO.checkProjectConfigByProjectId(projectId)){
+		 	projectConfigDAO.deleteByProjectId(projectId);
+		}
 	}
 	
 	private Result<Void> checkConfigParam(ProjectConfigDTO configDTO) {
